+++ src/Obsolete/LinearAlgebraUtilities.jl
