+++ src/Obsolete/MatrixSequence.jl
