+++ src/Obsolete/interference_based_sequence.jl
