+++ src/Obsolete/symp.jl
