+++ src/Obsolete/symp_generators.jl
